zmiany do sugesti z PL style -> class link pod stałą BASE_URL jeden service do wszystkiego
