+++ rental-angular/src/app/components/edit-company-data/edit-company-data.css
@@ -1,0 +1,10 @@
+
+.tdName{
+  width: 20%
+}
+.td57{
+  width: 57%
+}
+.tableDiv{
+  width: 100%
+}
