+++ rental-angular/src/app/components/edit-company-data/edit-company-data.html
@@ -6,7 +6,7 @@
     <div class="editblock">
 
       <h4>Edycja</h4>
-      <table style="width: 100%; height: 13em">
+      <table class="table100-13">
         <thead>
         <tr>
           <th ></th>
@@ -17,9 +17,9 @@
         </thead>
         <tbody>
         <tr>
-          <td style="width: 20%">    <label for="company_name">Nazwa:</label></td>
-          <td style="width: 57%">
-            <div style="width: 100%">
+          <td  class="tdName" >    <label for="company_name">Nazwa:</label></td>
+          <td class="td57">
+            <div class="tableDiv">
               <input type="text"  id="company_name" ></div></td>
           <td>
             <button class="button button-green"><i class="icon-wrench"></i></button>
@@ -31,7 +31,7 @@
             <label for="domain">Domena:</label>
 
           <td>
-            <div style="width: 100%"> <input type="text" id="domain" ></div></td>
+            <div class="tableDiv"> <input type="text" id="domain" ></div></td>
           <td>
             <button class="button button-green"><i class="icon-wrench"></i></button>
             <button  class="button button-red"><i class="icon-cancel-circled"></i></button>
@@ -40,7 +40,7 @@
         <tr>
           <td>      <label for="company_adress">Adres:</label>
           </td>
-          <td><div style="width: 100%"><input type="text" id="company_adress" ></div></td>
+          <td><div class="tableDiv"><input type="text" id="company_adress" ></div></td>
           <td>
             <button class="button button-green"><i class="icon-wrench"></i></button>
             <button  class="button button-red"><i class="icon-cancel-circled"></i></button>
@@ -49,7 +49,7 @@
         <tr>
           <td>      <label for="company_owner">Właściciel:</label>
           </td>
-          <td><div style="width: 100%"> <input type="text" id="company_owner" ></div></td>
+          <td><div class="tableDiv"> <input type="text" id="company_owner" ></div></td>
           <td>
             <button class="button button-green"><i class="icon-wrench"></i></button>
             <button  class="button button-red"><i class="icon-cancel-circled"></i></button>
@@ -59,7 +59,7 @@
           <td>         <label for="logotype">Logotyp</label>
 
           </td>
-          <td> <div style="width: 100%"> <input type="text" id="logotype" ></div></td>
+          <td> <div class="tableDiv"> <input type="text" id="logotype" ></div></td>
           <td>
             <button class="button button-green"><i class="icon-wrench"></i></button>
             <button  class="button button-red"><i class="icon-cancel-circled"></i></button>
