+++ rental-angular/src/app/components/home/home.component.css
@@ -1,3 +1,6 @@
 .editblock{
 min-width: 600px;
 }
+.tdWidth20{
+  width: 20%
+}
