+++ rental-angular/src/app/components/home/home.component.html
@@ -1,7 +1,7 @@
 <section>
   <div class="editblock">
     <h4>Dane firmy</h4>
-    <table style="width: 100%; height: 13em">
+    <table class="table100-13">
       <thead>
       <tr>
         <th> </th>
@@ -10,15 +10,15 @@
       </thead>
       <tbody>
       <tr>
-        <td style="width: 20%">     <label >Nazwa:</label></td>
-        <td><p>{{companyData.name}}</p>
+        <td class="tdWidth20">     <label >Nazwa:</label></td>
+        <td><p>{{companyData?.name}}</p>
         </td>
       </tr>
       <tr>
         <td>
           <label>Domena:</label>
 
-        <td>   <p>{{companyData.webDomain}}</p>     <td>
+        <td>   <p><a href="{{companyData?.webDomain}}">{{companyData?.webDomain}}</a></p>     <td>
       </td>
       </tr>
       <tr>
@@ -29,13 +29,13 @@
       <tr>
         <td>      <label>Właściciel:</label>
         </td>
-        <td> <p>{{companyData.owner}}</p></td>
+        <td> <p>{{companyData?.owner}}</p></td>
       </tr>
       <tr>
         <td>         <label >Logotyp</label>
 
         </td>
-        <td>  <p>{{companyData.logoURL}}</p></td>
+        <td>  <p>{{companyData?.logoURL}}</p></td>
       </tr>
       </tbody>
     </table>
