+++ rental-angular/src/app/components/home/home.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import {CompanyData, HomeService} from "../../services/home.service";
+import {CompanyData} from "../../service/company-data";
+import {RestService} from "../../service/rest.service";
 
 
 
@@ -13,7 +14,8 @@ export class HomeComponent implements OnInit {
 
 companyData : CompanyData;
 
-  constructor(private homeService : HomeService) { }
+
+  constructor(private homeService : RestService) { }
 
   ngOnInit(): void {
     this.homeService.getCompanyData().subscribe(value => {
