+++ rental-angular/src/styles.css
@@ -98,7 +98,11 @@ app-root {
   font-size: 25px;
   flex: auto;
 }
-/* Linki */
+.table100-13{
+  width: 100%;
+  height: 13em
+}
+
 .a-menu {
   padding-left: 10px;
   padding-right: 10px;
